db_server/requirements.txt
fastapi==0.95.2
starlette==0.27.0
httpx==0.19.0
psycopg2==2.9.7
langchain==0.0.220
python-dotenv==0.21.0
uvicorn==0.18.3
openai